+++ perl/FSM/Support/ISFPublicInterfaceContract.pm
@@ -532,6 +532,7 @@ sub build_isf_public_interface_contract {
             't/1345-isf-actor-param-actor-constants.t',
             't/1346-isf-actor-param-actor-params.t',
             't/1347-isf-transaction-param-actor-static-defaults.t',
+            't/1348-isf-transaction-param-transaction-params.t',
         ],
         guidance => [
             'Treat this as the first bounded public ISF downstream-consumer contract, advertised through embedding.isf_public_interface.',